crew_ai/pyproject.toml
[project]
name = "latest_ai_development"
version = "0.1.0"
description = "latest-ai-development using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.6.0",
    
    # Core
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # Document Processing
    "pymupdf>=1.23.0",
    "beautifulsoup4>=4.12.0",
    
    # Web Scraping & HTTP
    "requests>=2.31.0",
    
    # API Server
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "litellm>=1.0.0",
]

[project.scripts]
latest_ai_development = "latest_ai_development.main:run"
run_crew = "latest_ai_development.main:run"
train = "latest_ai_development.main:train"
replay = "latest_ai_development.main:replay"
test = "latest_ai_development.main:test"
run_with_trigger = "latest_ai_development.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
